+++ aoe4/aoe4_build_order.py
@@ -2,7 +2,7 @@ from aoe4.aoe4_civ_icon import aoe4_civilization_icon
 from common.build_order_tools import is_valid_resource
 
 
-def check_valid_aoe4_build_order(data: dict) -> bool:
+def check_valid_aoe4_build_order(data: dict) -> (bool, str):
     """Check if a build order is valid for AoE4
 
     Parameters
@@ -12,78 +12,103 @@ def check_valid_aoe4_build_order(data: dict) -> bool:
     Returns
     -------
     True if valid build order, False otherwise
+    string indicating the error (empty if no error)
     """
+    bo_name_str: str = ''
     try:
-        civilization_data = data['civilization']
         name: str = data['name']
+        bo_name_str: str = f'{name} | '
+
+        civilization_data = data['civilization']
         build_order: list = data['build_order']
 
         # check correct civilization
         if isinstance(civilization_data, list):  # list of civilizations
             if len(civilization_data) == 0:
-                print('Valid civilization list empty.')
-                return False
+                return False, bo_name_str + 'Valid civilization list is empty.'
 
             for civilization in civilization_data:
                 if civilization not in aoe4_civilization_icon:
-                    print(f'Unknown civilization \'{civilization}\' (check spelling) for build order \'{name}\'.')
-                    return False
+                    return False, bo_name_str + f'Unknown civilization \'{civilization}\' (check spelling).'
         elif civilization_data not in aoe4_civilization_icon:  # single civilization provided
-            print(f'Unknown civilization \'{civilization_data}\' (check spelling) for build order \'{name}\'.')
-            return False
+            return False, bo_name_str + f'Unknown civilization \'{civilization_data}\' (check spelling).'
 
-        count = len(build_order)  # size of the build order
-        if count < 1:
-            print(f'Build order \'{name}\' is empty.')
-            return False
+        if len(build_order) < 1:  # size of the build order
+            return False, bo_name_str + f'Build order is empty.'
 
         # loop on the build order steps
-        for item in build_order:
-            # check main fields are there
-            if ('population_count' not in item) or ('villager_count' not in item) or ('age' not in item) or (
-                    'resources' not in item) or ('notes' not in item):
-                print(f'Build order \'{name}\' does not have all the required fields.')
-                return False
+        for step_id, item in enumerate(build_order):
+            step_str = f'Step {step_id}'
+
+            # check if main fields are there
+            if 'population_count' not in item:
+                return False, bo_name_str + f'{step_str} is missing the \'population_count\' field.'
+
+            if 'villager_count' not in item:
+                return False, bo_name_str + f'{step_str} is missing the \'villager_count\' field.'
+
+            if 'age' not in item:
+                return False, bo_name_str + f'{step_str} is missing the \'age\' field.'
+
+            if 'resources' not in item:
+                return False, bo_name_str + f'{step_str} is missing the \'resources\' field.'
+
+            if 'notes' not in item:
+                return False, bo_name_str + f'{step_str} is missing the \'notes\' field.'
 
             # population count
             if not isinstance(item['population_count'], int):
-                print(f'Build order \'{name}\' does not have a valid population count.')
-                return False
+                return False, bo_name_str + f'{step_str} has invalid population count ({item["population_count"]}).'
 
             # villager count
             if not isinstance(item['villager_count'], int):
-                print(f'Build order \'{name}\' does not have a valid villager count.')
-                return False
+                return False, bo_name_str + f'{step_str} has invalid villager count ({item["villager_count"]}).'
 
             # age
             if (not isinstance(item['age'], int)) or (int(item['age']) > 4):
-                print(f'Build order \'{name}\' does not have a valid age number.')
-                return False
+                return False, bo_name_str + f'{step_str} has invalid age number ({item["age"]}) (max: 4 for Imperial).'
 
             # resources
             resources = item['resources']
-            if ('wood' not in resources) or ('food' not in resources) or ('gold' not in resources) or (
-                    'stone' not in resources):
-                print(f'Build order \'{name}\' does not have all the resources fields.')
-                return False
 
-            if (not is_valid_resource(resources['wood'])) or (not is_valid_resource(resources['food'])) or (
-                    not is_valid_resource(resources['gold'])) or (not is_valid_resource(resources['stone'])):
-                print(f'Build order \'{name}\' resources are not valid.')
-                return False
+            if 'wood' not in resources:
+                return False, bo_name_str + f'{step_str} is missing the \'wood\' field in \'resources\'.'
+
+            if 'food' not in resources:
+                return False, bo_name_str + f'{step_str} is missing the \'food\' field in \'resources\'.'
+
+            if 'gold' not in resources:
+                return False, bo_name_str + f'{step_str} is missing the \'gold\' field in \'resources\'.'
+
+            if 'stone' not in resources:
+                return False, bo_name_str + f'{step_str} is missing the \'stone\' field in \'resources\'.'
+
+            if not is_valid_resource(resources['wood']):
+                return False, bo_name_str + f'{step_str} has an invalid \'wood\' resource ({resources["wood"]}).'
+
+            if not is_valid_resource(resources['food']):
+                return False, bo_name_str + f'{step_str} has an invalid \'food\' resource ({resources["food"]}).'
+
+            if not is_valid_resource(resources['gold']):
+                return False, bo_name_str + f'{step_str} has an invalid \'gold\' resource ({resources["gold"]}).'
+
+            if not is_valid_resource(resources['stone']):
+                return False, bo_name_str + f'{step_str} has an invalid \'stone\' resource ({resources["stone"]}).'
 
             # optional builder count
             if ('builder' in resources) and (not is_valid_resource(resources['builder'])):
-                print(f'Build order \'{name}\' builder resource is not valid.')
-                return False
+                return False, bo_name_str + f'{step_str} has an invalid \'builder\' resource.'
 
             # notes
             notes = item['notes']
             for note in notes:
                 if not isinstance(note, str):
-                    print(f'Build order \'{name}\' contains wrong notes.')
-                    return False
-        return True
-    except KeyError:
-        print('Wrong key detected when checking for valid AoE4 build order.')
-        return False
+                    return False, bo_name_str + f'{step_str} note \'{note}\' is not a string.'
+
+    except KeyError as err:
+        return False, bo_name_str + f'Wrong JSON key: {err}.'
+
+    except Exception as err:
+        return False, bo_name_str + str(err)
+
+    return True, ''  # valid build order, no error message